wip: make new message component and mark one as "old"
