+++ packages/chat/src/components/Message/Message.component.tsx
@@ -1,0 +1,7 @@
+export const Message = () => {
+  return (
+    <div>
+      <h1>Message</h1>
+    </div>
+  );
+};
