+++ packages/chat/src/components/OldMessage/styled.ts
